chore: Refactor import path for Providers component
