+++ app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "@styles/globals.css";
-import Providers from "@libraries/providers";
+import Providers from "../core/components/providers";
 
 const inter = Inter({ subsets: ["latin"] });
